[BUGFIX] Layout problems with color picker
The jscolor library, which provides the color picker in the last step of
the installation process has some magic, which tries to find the best
position (above/below) the color value input field.

This magic fails in our use case, when the browser window is too small.
Our use case, to always show the picker is not really desired by the
jscolor library.
However, through a new option "pickerSmartPosition", which is included
in the latest jscolor release 1.3.11 after our request, we can disable
this magic positioning.

Change-Id: Idb092937ab4399df6975ef1227d9b69d46a76092
Resolves: #31694