[!!!][FEATURE] Replace EXT:tt_news with EXT:news
EXT:tt_news is very old and is updated to a more modern version.

The TypoScript is adopted to reflect the changes. The default
template of EXT:news has is copied to the fileadmin and is
altered to match the old template - although there are a few
minor differences.

Change-Id: Ieffd340dc016b6431345491384600920b4e34c72
Resolves: #39724
Releases: 6.0