[BUGFIX] Installer was broken due to central system environment check
With the commit 009f307bd2e1aca577f82044800286529ebaa5e5 the function isGD
was removed. This function is still used during the installation process
and raises a fatal error. Therefore the function calls have to be removed
and the installation of gd library has to be approved with proper
functions.

Fixes: #46515
Releases: 6.1

Change-Id: Id7c81b9ec9d2262c13418a30c1ceccbc86f0a884
Reviewed-on: https://review.typo3.org/19157
Reviewed-by: Alexander Opitz
Reviewed-by: Tom Ruether
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn